# docs/swagger.yaml
definitions:
...
  server.twoFactorCodeRequest:
    properties:
      code:
        type: string
    type: object
  server.twoFactorEnableRequest:
    properties:
      current_password:
        type: string
    type: object
  server.twoFactorPasswordRequest:
    properties:
      current_password:
        type: string
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: REST API aplikasi EBITDA Max APN (refactor Go + Next.js) — scope manager
    KDKMP & superadmin.
  title: EBITDA Max APN API
  version: "1.0"
paths:
...
  /api/v1/auth/two-factor-challenge:
    post:
      consumes:
      - application/json
      description: Menyelesaikan login yang tertahan 2FA memakai kode TOTP atau recovery
        code.
      parameters:
      - description: Kode TOTP atau recovery code
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.twoFactorCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verifikasi 2FA saat login
      tags:
      - Auth
...
  /api/v1/two-factor:
    delete:
      consumes:
      - application/json
      description: Menonaktifkan 2FA dan menghapus secret + recovery codes. Wajib
        kata sandi saat ini.
      parameters:
      - description: Kata sandi saat ini
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.twoFactorPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Nonaktifkan 2FA
      tags:
      - Two-Factor
    get:
      description: Mengetahui apakah 2FA aktif untuk user yang sedang login.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Status 2FA
      tags:
      - Two-Factor
  /api/v1/two-factor/confirm:
    post:
      consumes:
      - application/json
      description: Memverifikasi kode TOTP pertama dan mengaktifkan 2FA; mengembalikan
        recovery codes (tampil sekali).
      parameters:
      - description: Kode TOTP
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.twoFactorCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Konfirmasi 2FA
      tags:
      - Two-Factor
  /api/v1/two-factor/enable:
    post:
      consumes:
      - application/json
      description: Membuat secret TOTP baru (belum aktif sampai dikonfirmasi kode).
        Wajib kata sandi saat ini.
      parameters:
      - description: Kata sandi saat ini
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.twoFactorEnableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Mulai aktifkan 2FA
      tags:
      - Two-Factor
  /api/v1/two-factor/recovery-codes:
    post:
      consumes:
      - application/json
      description: Mengganti seluruh recovery codes 2FA. Wajib kata sandi saat ini.
      parameters:
      - description: Kata sandi saat ini
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.twoFactorPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Buat ulang recovery codes
      tags:
      - Two-Factor
...
schemes:
- http
securityDefinitions:
  CookieAuth:
    in: cookie
    name: ebitda_session
    type: apiKey
swagger: "2.0"